borg: react when banishment is the wonder effect
